Data.JSString

Description

Manipulation of JavaScript strings, API and fusion implementation based on Data.Text by Tom Harper, Duncan Coutts, Bryan O'Sullivan e.a.

Documentation

data JSString

A wrapper around a JavaScript string

Creation and elimination

pack :: String -> JSString

O(n) Convert a String into a JSString. Subject to fusion.

unpack :: JSString -> String

O(n) Convert a JSString into a String. Subject to fusion.

unpack' :: JSString -> String

singleton :: Char -> JSString

O(1) Convert a character into a JSString. Subject to fusion. Performs replacement on invalid scalar values.

empty :: JSString

O(1) The empty JSString.

Basic interface

cons :: Char -> JSString -> JSString infixr 5

O(n) Adds a character to the front of a JSString. This function is more costly than its List counterpart because it requires copying a new array. Subject to fusion. Performs replacement on invalid scalar values.

snoc :: JSString -> Char -> JSString

O(n) Adds a character to the end of a JSString. This copies the entire array in the process, unless fused. Subject to fusion. Performs replacement on invalid scalar values.

append :: JSString -> JSString -> JSString

O(n) Appends one JSString to the other by copying both of them into a new JSString. Subject to fusion.

uncons :: JSString -> Maybe (Char, JSString)

O(1) Returns the first character and rest of a JSString, or Nothing if empty. Subject to fusion.

head :: JSString -> Char

O(1) Returns the first character of a JSString, which must be non-empty. Subject to fusion.

last :: JSString -> Char

O(1) Returns the last character of a JSString, which must be non-empty. Subject to fusion.

tail :: JSString -> JSString

O(1) Returns all characters after the head of a JSString, which must be non-empty. Subject to fusion.

init :: JSString -> JSString

O(1) Returns all but the last character of a JSString, which must be non-empty. Subject to fusion.

null :: JSString -> Bool

O(1) Tests whether a JSString is empty or not. Subject to fusion.

length :: JSString -> Int

O(n) Returns the number of characters in a JSString. Subject to fusion.

compareLength :: JSString -> Int -> Ordering

O(n) Compare the count of characters in a JSString to a number. Subject to fusion.

This function gives the same answer as comparing against the result of length, but can short circuit if the count of characters is greater than the number, and hence be more efficient.

Transformations

map :: (Char -> Char) -> JSString -> JSString

O(n) map f t is the JSString obtained by applying f to each element of t. Subject to fusion. Performs replacement on invalid scalar values.

intercalate :: JSString -> [JSString] -> JSString

O(n) The intercalate function takes a JSString and a list of JSStrings and concatenates the list after interspersing the first argument between each element of the list.

intersperse :: Char -> JSString -> JSString

O(n) The intersperse function takes a character and places it between the characters of a JSString. Subject to fusion. Performs replacement on invalid scalar values.

transpose :: [JSString] -> [JSString]

O(n) The transpose function transposes the rows and columns of its JSString argument. Note that this function uses pack, unpack, and the list version of transpose, and is thus not very efficient.

reverse :: JSString -> JSString

O(n) Reverse the characters of a string. Subject to fusion.

replace

Arguments

:: JSString
needle to search for. If this string is empty, an error will occur.

-> JSString
replacement to replace needle with.

-> JSString
haystack in which to search.

-> JSString

O(m+n) Replace every non-overlapping occurrence of needle in haystack with replacement.

This function behaves as though it was defined as follows:

replace needle replacement haystack =
  intercalate replacement (splitOn needle haystack)

As this suggests, each occurrence is replaced exactly once. So if needle occurs in replacement, that occurrence will not itself be replaced recursively:

replace "oo" "foo" "oo" == "foo"

In cases where several instances of needle overlap, only the first one will be replaced:

replace "ofo" "bar" "ofofo" == "barfo"

In (unlikely) bad cases, this function's time complexity degrades towards O(n*m).

Case conversion

toCaseFold :: JSString -> JSString

O(n) Convert a string to folded case. Subject to fusion.

This function is mainly useful for performing caseless (also known as case insensitive) string comparisons.

A string x is a caseless match for a string y if and only if:

toCaseFold x == toCaseFold y

The result string may be longer than the input string, and may differ from applying toLower to the input string. For instance, the Armenian small ligature "ﬓ" (men now, U+FB13) is case folded to the sequence "մ" (men, U+0574) followed by "ն" (now, U+0576), while the Greek "µ" (micro sign, U+00B5) is case folded to "μ" (small letter mu, U+03BC) instead of itself.

toLower :: JSString -> JSString

O(n) Convert a string to lower case, using simple case conversion. Subject to fusion.

The result string may be longer than the input string. For instance, "İ" (Latin capital letter I with dot above, U+0130) maps to the sequence "i" (Latin small letter i, U+0069) followed by " ̇" (combining dot above, U+0307).

toUpper :: JSString -> JSString

O(n) Convert a string to upper case, using simple case conversion. Subject to fusion.

The result string may be longer than the input string. For instance, the German "ß" (eszett, U+00DF) maps to the two-letter sequence "SS".

toTitle :: JSString -> JSString

O(n) Convert a string to title case, using simple case conversion. Subject to fusion.

The first letter of the input is converted to title case, as is every subsequent letter that immediately follows a non-letter. Every letter that immediately follows another letter is converted to lower case.

The result string may be longer than the input string. For example, the Latin small ligature ﬂ (U+FB02) is converted to the sequence Latin capital letter F (U+0046) followed by Latin small letter l (U+006C).

Note: this function does not take language or culture specific rules into account. For instance, in English, different style guides disagree on whether the book name "The Hill of the Red Fox" is correctly title cased—but this function will capitalize every word.

Justification

justifyLeft :: Int -> Char -> JSString -> JSString

O(n) Left-justify a string to the given length, using the specified fill character on the right. Subject to fusion. Performs replacement on invalid scalar values.

Examples:

justifyLeft 7 'x' "foo"    == "fooxxxx"
justifyLeft 3 'x' "foobar" == "foobar"

justifyRight :: Int -> Char -> JSString -> JSString

O(n) Right-justify a string to the given length, using the specified fill character on the left. Performs replacement on invalid scalar values.

Examples:

justifyRight 7 'x' "bar"    == "xxxxbar"
justifyRight 3 'x' "foobar" == "foobar"

center :: Int -> Char -> JSString -> JSString

O(n) Center a string to the given length, using the specified fill character on either side. Performs replacement on invalid scalar values.

Examples:

center 8 'x' "HS" = "xxxHSxxx"

Folds

foldl :: (a -> Char -> a) -> a -> JSString -> a

O(n) foldl, applied to a binary operator, a starting value (typically the left-identity of the operator), and a JSString, reduces the JSString using the binary operator, from left to right. Subject to fusion.

foldl' :: (a -> Char -> a) -> a -> JSString -> a

O(n) A strict version of foldl. Subject to fusion.

foldl1 :: (Char -> Char -> Char) -> JSString -> Char

O(n) A variant of foldl that has no starting value argument, and thus must be applied to a non-empty JSString. Subject to fusion.

foldl1' :: (Char -> Char -> Char) -> JSString -> Char

O(n) A strict version of foldl1. Subject to fusion.

foldr :: (Char -> a -> a) -> a -> JSString -> a

O(n) foldr, applied to a binary operator, a starting value (typically the right-identity of the operator), and a JSString, reduces the JSString using the binary operator, from right to left. Subject to fusion.

foldr1 :: (Char -> Char -> Char) -> JSString -> Char

O(n) A variant of foldr that has no starting value argument, and thus must be applied to a non-empty JSString. Subject to fusion.

Special folds

concat :: [JSString] -> JSString

O(n) Concatenate a list of JSStrings.

concatMap :: (Char -> JSString) -> JSString -> JSString

O(n) Map a function over a JSString that results in a JSString, and concatenate the results.

any :: (Char -> Bool) -> JSString -> Bool

O(n) any p t determines whether any character in the JSString t satisifes the predicate p. Subject to fusion.

all :: (Char -> Bool) -> JSString -> Bool

O(n) all p t determines whether all characters in the JSString t satisify the predicate p. Subject to fusion.

maximum :: JSString -> Char

O(n) maximum returns the maximum value from a JSString, which must be non-empty. Subject to fusion.

minimum :: JSString -> Char

O(n) minimum returns the minimum value from a JSString, which must be non-empty. Subject to fusion.

Construction

Scans

scanl :: (Char -> Char -> Char) -> Char -> JSString -> JSString

O(n) scanl is similar to foldl, but returns a list of successive reduced values from the left. Subject to fusion. Performs replacement on invalid scalar values.

scanl f z [x1, x2, ...] == [z, z `f` x1, (z `f` x1) `f` x2, ...]

Note that

last (scanl f z xs) == foldl f z xs.

scanl1 :: (Char -> Char -> Char) -> JSString -> JSString

O(n) scanl1 is a variant of scanl that has no starting value argument. Subject to fusion. Performs replacement on invalid scalar values.

scanl1 f [x1, x2, ...] == [x1, x1 `f` x2, ...]

scanr :: (Char -> Char -> Char) -> Char -> JSString -> JSString

O(n) scanr is the right-to-left dual of scanl. Performs replacement on invalid scalar values.

scanr f v == reverse . scanl (flip f) v . reverse

scanr1 :: (Char -> Char -> Char) -> JSString -> JSString

O(n) scanr1 is a variant of scanr that has no starting value argument. Subject to fusion. Performs replacement on invalid scalar values.

Accumulating maps

mapAccumL :: (a -> Char -> (a, Char)) -> a -> JSString -> (a, JSString)

O(n) Like a combination of map and foldl'. Applies a function to each element of a JSString, passing an accumulating parameter from left to right, and returns a final JSString. Performs replacement on invalid scalar values.

mapAccumR :: (a -> Char -> (a, Char)) -> a -> JSString -> (a, JSString)

The mapAccumR function behaves like a combination of map and a strict foldr; it applies a function to each element of a JSString, passing an accumulating parameter from right to left, and returning a final value of this accumulator together with the new JSString. Performs replacement on invalid scalar values.

Generation and unfolding

replicate :: Int -> JSString -> JSString

O(n*m) replicate n t is a JSString consisting of the input t repeated n times.

unfoldr :: (a -> Maybe (Char, a)) -> a -> JSString

O(n), where n is the length of the result. The unfoldr function is analogous to the List unfoldr. unfoldr builds a JSString from a seed value. The function takes the element and returns Nothing if it is done producing the JSString, otherwise Just (a,b). In this case, a is the next Char in the string, and b is the seed value for further production. Subject to fusion. Performs replacement on invalid scalar values.

unfoldrN :: Int -> (a -> Maybe (Char, a)) -> a -> JSString

O(n) Like unfoldr, unfoldrN builds a JSString from a seed value. However, the length of the result should be limited by the first argument to unfoldrN. This function is more efficient than unfoldr when the maximum length of the result is known and correct, otherwise its performance is similar to unfoldr. Subject to fusion. Performs replacement on invalid scalar values.

Substrings

Breaking strings

take :: Int -> JSString -> JSString

O(n) take n, applied to a JSString, returns the prefix of the JSString of length n, or the JSString itself if n is greater than the length of the JSString. Subject to fusion.

takeEnd :: Int -> JSString -> JSString

O(n) takeEnd n t returns the suffix remaining after taking n characters from the end of t.

Examples:

takeEnd 3 "foobar" == "bar"

drop :: Int -> JSString -> JSString

O(n) drop n, applied to a JSString, returns the suffix of the JSString after the first n characters, or the empty JSString if n is greater than the length of the JSString. Subject to fusion.

dropEnd :: Int -> JSString -> JSString

O(n) dropEnd n t returns the prefix remaining after dropping n characters from the end of t.

Examples:

dropEnd 3 "foobar" == "foo"

takeWhile :: (Char -> Bool) -> JSString -> JSString

O(n) takeWhile, applied to a predicate p and a JSString, returns the longest prefix (possibly empty) of elements that satisfy p. Subject to fusion.

dropWhile :: (Char -> Bool) -> JSString -> JSString

O(n) dropWhile p t returns the suffix remaining after takeWhile p t. Subject to fusion.

dropWhileEnd :: (Char -> Bool) -> JSString -> JSString

O(n) dropWhileEnd p t returns the prefix remaining after dropping characters that fail the predicate p from the end of t. Subject to fusion. Examples:

dropWhileEnd (=='.') "foo..." == "foo"

dropAround :: (Char -> Bool) -> JSString -> JSString

O(n) dropAround p t returns the substring remaining after dropping characters that fail the predicate p from both the beginning and end of t. Subject to fusion.

strip :: JSString -> JSString

O(n) Remove leading and trailing white space from a string. Equivalent to:

dropAround isSpace

stripStart :: JSString -> JSString

O(n) Remove leading white space from a string. Equivalent to:

dropWhile isSpace

stripEnd :: JSString -> JSString

O(n) Remove trailing white space from a string. Equivalent to:

dropWhileEnd isSpace

splitAt :: Int -> JSString -> (JSString, JSString)

O(n) splitAt n t returns a pair whose first element is a prefix of t of length n, and whose second is the remainder of the string. It is equivalent to (take n t, drop n t).

breakOn :: JSString -> JSString -> (JSString, JSString)

O(n+m) Find the first instance of needle (which must be non-null) in haystack. The first element of the returned tuple is the prefix of haystack before needle is matched. The second is the remainder of haystack, starting with the match.

Examples:

breakOn "::" "a::b::c" ==> ("a", "::b::c")
breakOn "/" "foobar"   ==> ("foobar", "")

Laws:

append prefix match == haystack
  where (prefix, match) = breakOn needle haystack

If you need to break a string by a substring repeatedly (e.g. you want to break on every instance of a substring), use breakOnAll instead, as it has lower startup overhead.

In (unlikely) bad cases, this function's time complexity degrades towards O(n*m).

breakOnEnd :: JSString -> JSString -> (JSString, JSString)

O(n+m) Similar to breakOn, but searches from the end of the string.

The first element of the returned tuple is the prefix of haystack up to and including the last match of needle. The second is the remainder of haystack, following the match.

breakOnEnd "::" "a::b::c" ==> ("a::b::", "c")

break :: (Char -> Bool) -> JSString -> (JSString, JSString)

O(n) break is like span, but the prefix returned is over elements that fail the predicate p.

span :: (Char -> Bool) -> JSString -> (JSString, JSString)

O(n) span, applied to a predicate p and text t, returns a pair whose first element is the longest prefix (possibly empty) of t of elements that satisfy p, and whose second is the remainder of the list.

group :: JSString -> [JSString]

O(n) Group characters in a string by equality.

group' :: JSString -> [JSString]

groupBy :: (Char -> Char -> Bool) -> JSString -> [JSString]

O(n) Group characters in a string according to a predicate.

inits :: JSString -> [JSString]

O(n^2) Return all initial segments of the given JSString, shortest first.

tails :: JSString -> [JSString]

O(n^2) Return all final segments of the given JSString, longest first.

Breaking into many substrings

splitOn

Arguments

:: JSString
String to split on. If this string is empty, an error will occur.

-> JSString
Input text.

-> [JSString]

O(m+n) Break a JSString into pieces separated by the first JSString argument (which cannot be empty), consuming the delimiter. An empty delimiter is invalid, and will cause an error to be raised.

Examples:

splitOn "\r\n" "a\r\nb\r\nd\r\ne" == ["a","b","d","e"]
splitOn "aaa"  "aaaXaaaXaaaXaaa"  == ["","X","X","X",""]
splitOn "x"    "x"                == ["",""]

and

intercalate s . splitOn s         == id
splitOn (singleton c)             == split (==c)

(Note: the string s to split on above cannot be empty.)

In (unlikely) bad cases, this function's time complexity degrades towards O(n*m).

splitOn'

Arguments

:: JSString
String to split on. If this string is empty, an error will occur.

-> JSString
Input text.

-> [JSString]

split :: (Char -> Bool) -> JSString -> [JSString]

O(n) Splits a JSString into components delimited by separators, where the predicate returns True for a separator element. The resulting components do not contain the separators. Two adjacent separators result in an empty component in the output. eg.

split (=='a') "aabbaca" == ["","","bb","c",""]
split (=='a') ""        == [""]

chunksOf :: Int -> JSString -> [JSString]

O(n) Splits a JSString into components of length k. The last element may be shorter than the other chunks, depending on the length of the input. Examples:

chunksOf 3 "foobarbaz"   == ["foo","bar","baz"]
chunksOf 4 "haskell.org" == ["hask","ell.","org"]

chunksOf' :: Int -> JSString -> [JSString]

O(n) Splits a JSString into components of length k. The last element may be shorter than the other chunks, depending on the length of the input. Examples:

chunksOf 3 "foobarbaz"   == ["foo","bar","baz"]
chunksOf 4 "haskell.org" == ["hask","ell.","org"]

Breaking into lines and words

lines :: JSString -> [JSString]

O(n) Breaks a JSString up into a list of JSStrings at newline Chars. The resulting strings do not contain newlines.

lines' :: JSString -> [JSString]

words :: JSString -> [JSString]

O(n) Breaks a JSString up into a list of words, delimited by Chars representing white space.

words' :: JSString -> [JSString]

unlines :: [JSString] -> JSString

O(n) Joins lines, after appending a terminating newline to each.

unwords :: [JSString] -> JSString

O(n) Joins words using single space characters.

Predicates

isPrefixOf :: JSString -> JSString -> Bool

O(n) The isPrefixOf function takes two JSStrings and returns True iff the first is a prefix of the second. Subject to fusion.

isSuffixOf :: JSString -> JSString -> Bool

O(n) The isSuffixOf function takes two JSStrings and returns True iff the first is a suffix of the second.

isInfixOf :: JSString -> JSString -> Bool

The isInfixOf function takes two JSStrings and returns True iff the first is contained, wholly and intact, anywhere within the second.

Complexity depends on how the JavaScript engine implements String.prototype.find.

View patterns

stripPrefix :: JSString -> JSString -> Maybe JSString

O(n) Return the suffix of the second string if its prefix matches the entire first string.

Examples:

stripPrefix "foo" "foobar" == Just "bar"
stripPrefix ""    "baz"    == Just "baz"
stripPrefix "foo" "quux"   == Nothing

This is particularly useful with the ViewPatterns extension to GHC, as follows:

{-# LANGUAGE ViewPatterns #-}
import Data.Text as T

fnordLength :: JSString -> Int
fnordLength (stripPrefix "fnord" -> Just suf) = T.length suf
fnordLength _                                 = -1

stripSuffix :: JSString -> JSString -> Maybe JSString

O(n) Return the prefix of the second string if its suffix matches the entire first string.

Examples:

stripSuffix "bar" "foobar" == Just "foo"
stripSuffix ""    "baz"    == Just "baz"
stripSuffix "foo" "quux"   == Nothing

This is particularly useful with the ViewPatterns extension to GHC, as follows:

{-# LANGUAGE ViewPatterns #-}
import Data.Text as T

quuxLength :: Text -> Int
quuxLength (stripSuffix "quux" -> Just pre) = T.length pre
quuxLength _                                = -1

commonPrefixes :: JSString -> JSString -> Maybe (JSString, JSString, JSString)

O(n) Find the longest non-empty common prefix of two strings and return it, along with the suffixes of each string at which they no longer match.

If the strings do not have a common prefix or either one is empty, this function returns Nothing.

Examples:

commonPrefixes "foobar" "fooquux" == Just ("foo","bar","quux")
commonPrefixes "veeble" "fetzer"  == Nothing
commonPrefixes "" "baz"           == Nothing

Searching

filter :: (Char -> Bool) -> JSString -> JSString

O(n) filter, applied to a predicate and a JSString, returns a JSString containing those characters that satisfy the predicate.

breakOnAll

Arguments

:: JSString
needle to search for

-> JSString
haystack in which to search

-> [(JSString, JSString)]

O(n+m) Find all non-overlapping instances of needle in haystack. Each element of the returned list consists of a pair:

The entire string prior to the kth match (i.e. the prefix)

The kth match, followed by the remainder of the string

Examples:

breakOnAll "::" ""
==> []
breakOnAll "/" "a/b/c/"
==> [("a", "/b/c/"), ("a/b", "/c/"), ("a/b/c", "/")]

In (unlikely) bad cases, this function's time complexity degrades towards O(n*m).

The needle parameter may not be empty.

breakOnAll'

Arguments

:: JSString
needle to search for

-> JSString
haystack in which to search

-> [(JSString, JSString)]

find :: (Char -> Bool) -> JSString -> Maybe Char

O(n) The find function takes a predicate and a JSString, and returns the first element matching the predicate, or Nothing if there is no such element.

partition :: (Char -> Bool) -> JSString -> (JSString, JSString)

O(n) The partition function takes a predicate and a JSString, and returns the pair of JSStrings with elements which do and do not satisfy the predicate, respectively; i.e.

partition p t == (filter p t, filter (not . p) t)

Indexing

index :: JSString -> Int -> Char

O(n) JSString index (subscript) operator, starting from 0.

findIndex :: (Char -> Bool) -> JSString -> Maybe Int

O(n) The findIndex function takes a predicate and a JSString and returns the index of the first element in the JSString satisfying the predicate. Subject to fusion.

count :: JSString -> JSString -> Int

O(n+m) The count function returns the number of times the query string appears in the given JSString. An empty query string is invalid, and will cause an error to be raised.

In (unlikely) bad cases, this function's time complexity degrades towards O(n*m).

Zipping

zip :: JSString -> JSString -> [(Char, Char)]

O(n) zip takes two JSStrings and returns a list of corresponding pairs of bytes. If one input JSString is short, excess elements of the longer JSString are discarded. This is equivalent to a pair of unpack operations.

zipWith :: (Char -> Char -> Char) -> JSString -> JSString -> JSString

O(n) zipWith generalises zip by zipping with the function given as the first argument, instead of a tupling function. Performs replacement on invalid scalar values.