Data.JSString.Read

Documentation

isInteger :: JSString -> Bool

Returns whether the JSString represents an integer at base 10

isNatural :: JSString -> Bool

Returns whether the JSString represents a natural number at base 10 (including 0)

readInt :: JSString -> Int

Convert a JSString to an Int, throwing an exception if it cannot be converted. Leading spaces and trailing non-digit characters are not allowed.

readIntMaybe :: JSString -> Maybe Int

Convert a JSString to an Int, returning Nothing if it cannot be converted. Leading spaces and trailing non-digit characters are not allowed.

lenientReadInt :: JSString -> Int

Convert a JSString to an Int, throwing an exception if it cannot be converted. Leading spaces are allowed. The function ignores trailing non-digit characters.

readInt64 :: JSString -> Int64

readInt64Maybe :: JSString -> Maybe Int64

readWord64 :: JSString -> Word64

readWord64Maybe :: JSString -> Maybe Word64

readDouble :: JSString -> Double

Convert a JSString to an Int, throwing an exception if it cannot be converted. Leading spaces are allowed. The function ignores trailing non-digit characters.

readDoubleMaybe :: JSString -> Maybe Double

Convert a JSString to a Double, returning Nothing if it cannot be converted. Leading spaces are allowed. The function ignores trailing non-digit characters.

readInteger :: JSString -> Integer

readIntegerMaybe :: JSString -> Maybe Integer