GHCJS.Marshal.Internal

Documentation

class FromJSVal a where

Minimal complete definition

Nothing

Methods

fromJSVal :: JSVal -> IO (Maybe a)

fromJSValUnchecked :: JSVal -> IO a

fromJSValListOf :: JSVal -> IO (Maybe [a])

fromJSValUncheckedListOf :: JSVal -> IO [a]

class ToJSVal a where

Minimal complete definition

Nothing

Methods

toJSVal :: a -> IO JSVal

toJSValListOf :: [a] -> IO JSVal

class PToJSVal a where

Methods

pToJSVal :: a -> JSVal

class PFromJSVal a where

Methods

pFromJSVal :: JSVal -> a

data Purity

Constructors

PureShared
conversion is pure even if the original value is shared

PureExclusive
conversion is pure if the we only convert once

toJSVal_generic :: forall a. (Generic a, GToJSVal (Rep a ())) => (String -> String) -> a -> IO JSVal

fromJSVal_generic :: forall a. (Generic a, GFromJSVal (Rep a ())) => (String -> String) -> JSVal -> IO (Maybe a)

toJSVal_pure :: PToJSVal a => a -> IO JSVal

fromJSVal_pure :: PFromJSVal a => JSVal -> IO (Maybe a)

fromJSValUnchecked_pure :: PFromJSVal a => JSVal -> IO a