GHCJS.Types

Documentation

data JSVal :: *

data WouldBlockException :: *

If a synchronous thread tries to do something that can only be done asynchronously, and the thread is set up to not continue asynchronously, it receives this exception.

Constructors

WouldBlockException

data JSException :: *

Constructors

JSException JSVal String

class IsJSVal a

jsval :: IsJSVal a => a -> JSVal

isNull :: JSVal -> Bool

isUndefined :: JSVal -> Bool

nullRef :: JSVal

data JSString

A wrapper around a JavaScript string

mkRef :: ByteArray# -> JSVal

type Ref# = ByteArray#

toPtr :: JSVal -> Ptr a

fromPtr :: Ptr a -> JSVal

type JSRef a = JSVal

Deprecated: Use JSVal instead, or a more specific newtype wrapper of JSVal

This is a deprecated copmatibility wrapper for the old JSRef type.

See https://github.com/ghcjs/ghcjs/issues/421