JavaScript.JSON.Types.Instances

Description

Types for working with JSON data.

Type classes

Core JSON classes

class FromJSON a where

Minimal complete definition

Nothing

Methods

parseJSON :: Value -> Parser a

class ToJSON a where

Minimal complete definition

Nothing

Methods

toJSON :: a -> Value

Generic JSON classes

class GFromJSON f where

Methods

gParseJSON :: Options -> Value -> Parser (f a)

class GToJSON f where

Methods

gToJSON :: Options -> f a -> Value

genericToJSON :: (Generic a, GToJSON (Rep a)) => Options -> a -> Value

genericParseJSON :: (Generic a, GFromJSON (Rep a)) => Options -> Value -> Parser a

Types

newtype DotNetTime :: *

A newtype wrapper for UTCTime that uses the same non-standard serialization format as Microsoft .NET, whose System.DateTime type is by default serialized to JSON as in the following example:

/Date(1302547608878)/

The number represents milliseconds since the Unix epoch.

Constructors

DotNetTime

Fields

fromDotNetTime :: UTCTime

Acquire the underlying value.

Inspecting Values

withObject :: String -> (Object -> Parser a) -> Value -> Parser a

withObject expected f value applies f to the Object when value is an Object and fails using typeMismatch expected otherwise.

withJSString :: String -> (JSString -> Parser a) -> Value -> Parser a

withArray :: String -> (JSArray -> Parser a) -> Value -> Parser a

withArray expected f value applies f to the Array when value is an Array and fails using typeMismatch expected otherwise.

withDouble :: String -> (Double -> Parser a) -> Value -> Parser a

withScientific expected f value applies f to the Double number when value is a Number. and fails using typeMismatch expected otherwise.

withBool :: String -> (Bool -> Parser a) -> Value -> Parser a

withBool expected f value applies f to the Bool when value is a Bool and fails using typeMismatch expected otherwise.

Functions

fromJSON :: FromJSON a => Value -> Result a

Convert a value from JSON, failing if the types do not match.

(.:) :: FromJSON a => Object -> JSString -> Parser a

Retrieve the value associated with the given key of an Object. The result is empty if the key is not present or the value cannot be converted to the desired type.

This accessor is appropriate if the key and value must be present in an object for it to be valid. If the key and value are optional, use '(.:?)' instead.

(.:?) :: FromJSON a => Object -> JSString -> Parser (Maybe a)

Retrieve the value associated with the given key of an Object. The result is Nothing if the key is not present, or empty if the value cannot be converted to the desired type.

This accessor is most useful if the key and value can be absent from an object without affecting its validity. If the key and value are mandatory, use '(.:)' instead.

(.!=) :: Parser (Maybe a) -> a -> Parser a

Helper for use in combination with .:? to provide default values for optional JSON object fields.

This combinator is most useful if the key and value can be absent from an object without affecting its validity and we know a default value to assign in that case. If the key and value are mandatory, use '(.:)' instead.

Example usage:

 v1 <- o .:? "opt_field_with_dfl" .!= "default_val"
 v2 <- o .:  "mandatory_field"
 v3 <- o .:? "opt_field2"

(.=) :: ToJSON a => JSString -> a -> Pair

Construct a Pair from a key and a value.

typeMismatch

Arguments

:: String
The name of the type you are trying to parse.

-> Value
The actual value encountered.

-> Parser a

Fail parsing due to a type mismatch, with a descriptive message.