JavaScript.Object

Documentation

data Object

create :: IO Object

create an empty object

getProp :: JSString -> Object -> IO JSVal

get a property from an object. If accessing the property results in an exception, the exception is converted to a JSException. Since exception handling code prevents some optimizations in some JS engines, you may want to use unsafeGetProp instead

unsafeGetProp :: JSString -> Object -> IO JSVal

setProp :: JSString -> JSVal -> Object -> IO ()

unsafeSetProp :: JSString -> JSVal -> Object -> IO ()

allProps :: Object -> IO JSArray

listProps :: Object -> IO [JSString]

isInstanceOf :: Object -> JSVal -> Bool