JavaScript.Web.Performance

Description

The Performance interface represents timing-related performance information for the given page.

Documentation

now :: IO Double

The now computation returns a high resolution time stamp, measured in milliseconds, accurate to one thousandth of a millisecond.

The value represented by 0 varies according the context, but in dedicated workers created from a Window context, the epoch is the value of the PerformanceTiming.navigationStart property.